fix bug overflow due to convert ulong to int

+++ program C/experiment/ShiftBits.c
@@ -17,7 +17,7 @@ int getMax(unsigned long id){
     return ((id >> 32) & USHRT_MAX);
 }
 
-int increase(unsigned long id, int portENB){
+unsigned long increase(unsigned long id, int portENB){
     if(id == -1){
         id = 0;
         id |= 1;
